fix kern_tables
The variable [len] read at the beginning of a table corresponds to the
full length of the table, header included. The Microsoft spec [0]
says:

> Type		Field	Description
> uint16	length	Length of the subtable, in bytes (including this header).

So when reading the pairs, the number of bytes to be read is len minus
3 uint16 (version, length, coverage) and minus the 4 uint16 that are
skipped just before the pairs.

I also modified the termination condition accordingly so that it's
clearer we stop when all bytes are read.

[0] https://docs.microsoft.com/en-us/typography/opentype/spec/kern